On demand tasks working.
Also added in time and duration types for tasks.

After a demand the scheduler tries to schedule back in the previously scheduled but unexecuted tasks. If this is not successful then these tasks are dropped. If these are successfully scheduled back in then it also tries to schedule back in the task which was interrupted by the demand. If this is not possible only the interrupted task is dropped.

Demands can be interrupted by timeout and by subsequent demanded tasks.