Automatically set relaxed_nav parameter.
The problem we have is that early estimates of navigation durations can be low, causing navigation actions to be killed even when they are working. The relaxed_nav parameter greatly inflates the estimates to prevent this happening, but must be set manually.

This PR adds a node which automatically sets the value of the relaxed_nav based on the number of nav stats for each edge. This is a rather coarse way of doing it, but given the separation of concerns in the system there is not other way. The better future solution is to get some kind of confidence measure with the estimate.